+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+## v0.2.0
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+### Added
+
+- Added r-hat convergence test to the Gibbs sampling process.
+- Added convergence diagnosis plots
+
+### Changed
+
+### Fixed
 
 ## v0.1.32